Split out the Todos pages for users.
This module was very specific to todos and was getting pretty large.

Here, we've made somewhat of an exception for the liberal use of
re-exports.
My guess here is that these re-exports are benign in this case: the
importing modules needn't be aware of the user-pages structure.

@noteed thoughts?
